Move noise stuff to its own file + formatting and mypy cleanup
