+++ vsmuxtools/video/encoders/standalone.py
@@ -7,89 +7,22 @@ from muxtools.utils.env import get_binary_version
 from muxtools.utils.dataclass import dataclass, allow_extra
 import re
 import json
-import random
 import sys
 import os
-
 from .base import SupportsQP, VideoEncoder
 from .types import LosslessPreset
+from .noise import (
+    SVTAV1_LIGHT_NOISE_TABLE_FULL,
+    SVTAV1_LIGHT_NOISE_TABLE_LIMITED,
+    x265_write_light_noise_table_full,
+    x265_write_light_noise_table_limited,
+)
 from ..settings import shift_zones, zones_to_args, norm_zones
 from ..clip_metadata import props_dict, SVT_AV1_RANGES
 
 from vsmuxtools.utils.source import generate_svt_av1_keyframes, src_file
 
 __all__ = ["x264", "x265", "LosslessX264", "SVTAV1"]
-
-
-SVTAV1_LIGHT_NOISE_TABLE_LIMITED = """filmgrn1
-E 0 18446744073709551615 1 787 1
-	p 3 7 0 8 0 1 128 192 256 128 192 256
-    sY 9 0 0 16 0 17 2 18 3 157 3 177 4 233 4 235 0 255 0
-	sCb 0
-	sCr 0
-	cY 3 4 3 3 3 3 3 3 4 2 0 2 3 3 3 2 -7 -19 -4 1 3 2 0 -18
-	cCb -3 9 -15 20 -6 0 0 9 -22 32 -50 10 -3 1 -15 32 -61 70 -26 -1 -2 17 -40 59 11
-	cCr -3 9 -15 20 -6 0 1 9 -21 32 -50 10 -3 0 -14 31 -61 71 -26 -1 -1 17 -40 58 11
-"""
-"""
-A table for photon noise that serves as a light dither layer to prevent banding.\n
-With cutoffs for limited range clips.
-"""
-
-SVTAV1_LIGHT_NOISE_TABLE_FULL = """filmgrn1
-E 0 18446744073709551615 1 787 1
-	p 3 7 0 8 0 1 128 192 256 128 192 256
-	sY 6 0 4 20 3 157 3 177 4 235 4 255 5
-	sCb 0
-	sCr 0
-	cY 3 4 3 3 3 3 3 3 4 2 0 2 3 3 3 2 -7 -19 -4 1 3 2 0 -18
-	cCb -3 9 -15 20 -6 0 0 9 -22 32 -50 10 -3 1 -15 32 -61 70 -26 -1 -2 17 -40 59 11
-	cCr -3 9 -15 20 -6 0 1 9 -21 32 -50 10 -3 0 -14 31 -61 71 -26 -1 -1 17 -40 58 11
-"""
-"""
-A table for photon noise that serves as a light dither layer to prevent banding.\n
-With cutoffs for full range clips.
-"""
-
-def x265_write_light_noise_table_helper(target: PathLike, length: int, s: str):
-    seed_pool = [65506, 65501, 65484, 65476, 65466, 65464, 65420, 65417, 65391, 65345, 65333, 65299,
-                 65260, 64921, 64917, 64831, 64774, 64693, 64448, 64436, 64435, 64423, 64384, 64332,
-                 64285, 64274, 64240, 64189, 64176, 64126, 64113, 64093, 63947, 63647, 63580, 63507,
-                 63504, 63456, 63023, 62518, 62359, 62258, 62156, 62143, 62040, 61851, 61692, 61482,
-                 61476, 60973, 60878, 60711, 60619, 60584, 60537, 60501, 60123, 59991, 59929, 59846,
-                 59724, 59669, 59665, 59637, 59625, 59621, 59180, 59119]
-    seeds = [63504]
-    while(len(seeds) < length):
-        n = random.choice(seed_pool)
-        if n not in seeds[-32:]:
-            seeds.append(n)
-
-    with target.open("wb") as f:
-        for seed in seeds:
-            f.write((1).to_bytes(4, byteorder="little", signed=True))
-            f.write((seed).to_bytes(2, byteorder="little", signed=False))
-            f.write((1).to_bytes(4, byteorder="little", signed=True))
-        
-            for n in s.split():
-                f.write((int(n)).to_bytes(4, byteorder="little", signed=True))
-            f.write((0).to_bytes(4, byteorder="little", signed=True))
-            f.write((0).to_bytes(4, byteorder="little", signed=True))
-            f.write((8).to_bytes(4, byteorder="little", signed=True))
-        
-            f.write((3).to_bytes(4, byteorder="little", signed=True))
-            for n in "3 4 3 3 3 3 3 3 4 2 0 2 3 3 3 2 -7 -19 -4 1 3 2 0 -18".split():
-                f.write((int(n)).to_bytes(4, byteorder="little", signed=True))
-            f.write((7).to_bytes(4, byteorder="little", signed=True))
-            f.write((0).to_bytes(4, byteorder="little", signed=True))
-        
-            f.write((1).to_bytes(4, byteorder="little", signed=True))
-            f.write((1).to_bytes(4, byteorder="little", signed=True))
-
-def x265_write_light_noise_table_limited(target: PathLike, length: int):
-    x265_write_light_noise_table_helper(target, length, "9 0 0 16 0 17 2 18 3 157 3 177 4 233 4 235 0 255 0")
-
-def x265_write_light_noise_table_full(target: PathLike, length: int):
-    x265_write_light_noise_table_helper(target, length, "6 0 4 20 3 157 3 177 4 235 4 255 5")
 
 
 @dataclass(config=allow_extra)
@@ -202,6 +135,7 @@ class x265(SupportsQP):
             args.extend(["--csv", str(csv_file)])
         if qpfile:
             args.extend(["--qpfile", qpfile])
+
         if self.light_photon_noise:
             if not any(key in self.get_custom_args_dict() for key in {"aom_film_grain", "film_grain"}):
                 fgs_table = get_workdir() / "x265_grain.bin"
@@ -211,19 +145,21 @@ class x265(SupportsQP):
                     x265_write_light_noise_table_full(fgs_table, clip.num_frames)
                 try:
                     fgs_table = fgs_table.relative_to(Path.cwd())
-                    if (os.name == "nt" and sys.version_info[1] >= 12) or \
-                       (os.name == "posix" and sys.version_info[1] >= 7):
+                    if (os.name == "nt" and sys.version_info[1] >= 12) or (os.name == "posix" and sys.version_info[1] >= 7):
                         if Path.cwd().is_mount():
                             raise ValueError
                 # Although I didn't find the code for it, in the tests it looks like vs-muxtools will always switch cwd to workdir,
                 # so the `relative_to` will always success and this `except` will never trigger, which is nice.
                 except ValueError:
+                    # fmt: off
                     warn("Due to an oversight in x265, the complete path for the aom film grain table will be logged into the video stream, viewable using tools like MediaInfo.", self)
                     warn("vs-muxtools by default tries to make the path relative. This way video stream only contains the path from cwd to the film grain table in the workdir.", self)
                     warn("This warning is displayed either when workdir is not relative to cwd, or your cwd is a mount point.", self)
                     warn("Right now, this below is the path that will be logged. If this doesn't contain anything sensitive, everything is good. If it does, change your cwd and rerun the encode.", self)
                     warn(str(fgs_table), self)
+                    # fmt: on
                 self.update_custom_args(aom_film_grain=str(fgs_table))
+
         if self.settings:
             args.extend(self.settings if isinstance(self.settings, list) else shlex.split(str(self.settings)))
         if self.zones:
@@ -320,11 +256,7 @@ class SVTAV1(VideoEncoder):
                 warn(f"Encoder version expected by the settings_builder: {self._settings_builder_id}.", self, 2)
 
         if not self.sd_clip and not self._encoder_id.startswith("SVT-AV1-Essential") and "_c" not in self.get_custom_args_dict():
-            warn(
-                "Providing a clip or a file for scene detection is recommended for SVT-AV1.",
-                self,
-                2,
-            )
+            warn("Providing a clip or a file for scene detection is recommended for SVT-AV1.", self, 2)
 
     def encode(self, clip: vs.VideoNode, outfile: PathLike | None = None) -> VideoFile:
         if clip.format.bits_per_sample > 10:
